// src/api/model/vacancy.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface Vacancy {
  source?: string;
  externalId?: string;
  title?: string;
  company?: string;
  companyUrl?: string;
  city?: string;
  salaryFrom?: number;
  salaryTo?: number;
  currency?: string;
  description?: string;
  experienceReq?: string;
  employmentType?: string;
  schedule?: string;
  publishedAt?: string;
  url?: string;
}